RPI_codes/requirements.txt
pyserial>=3.5
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
geojson>=3.0.0
shapely>=2.0.0
RPi.GPIO>=0.7.1
gpiozero>=2.0.0
numpy>=1.24.0
requests>=2.31.0